Fix render-grid replay test helpers
