+++ Packages/Shared/CMUXMobileCore/Tests/CMUXMobileCoreTests/MobileTerminalRenderGridTests.swift
@@ -19,29 +19,28 @@ import Testing
 
     let decoded = try MobileTerminalRenderGridFrame.decodeJSONObject(frame.jsonObject())
     #expect(decoded == frame)
-    // A full snapshot is restored as a synchronized, autowrap-off scrolling
-    // flow: reset supported state, apply default colors/style, clear, paint each
-    // viewport row (CHA-positioned spans), then restore the cursor.
-    #expect(String(data: frame.vtReplacementBytes(), encoding: .utf8) ==
-        "\u{1B}[?2026h\u{1B}[0$}\u{1B}[>m\u{1B}[r\u{1B}[?69l\u{1B}[?5W\u{1B}[?1049l" +
-        "\u{1B}]8;;\u{1B}\\" +
-        "\u{1B}[1\"q\u{1B}[0\"q\u{1B}[999<u\u{1B}[0;1=u\u{0F}\u{1B}(B\u{1B})B\u{1B}*B\u{1B}+B" +
-        "\u{1B}]110\u{1B}\\\u{1B}]111\u{1B}\\\u{1B}]112\u{1B}\\" +
-        "\u{1B}[0m" +
-        "\u{1B}[H\u{1B}[2J\u{1B}[3J\u{1B}[?1049h" +
-        "\u{1B}]8;;\u{1B}\\" +
-        "\u{1B}[1\"q\u{1B}[0\"q\u{1B}[999<u\u{1B}[0;1=u\u{0F}\u{1B}(B\u{1B})B\u{1B}*B\u{1B}+B" +
-        "\u{1B}[0m" +
-        "\u{1B}[H\u{1B}[2J\u{1B}[?1049l\u{1B}[H" +
-        "\u{1B}[?7l\u{1B}[?25l\u{1B}[0m" +
-        "\u{1B}[0m\u{1B}[1Galpha" +
-        "\r\n\u{1B}[0m" +
-        "\r\n\u{1B}[0m\u{1B}[1G beta" +
-        "\r\n\u{1B}[0m" +
-        "\u{1B}[0m\u{1B}[2 q\u{1B}[?25h\u{1B}[3;6H" +
-        "\u{1B}7" +
-        "\u{1B}[?2026l"
-    )
+    let actual = String(data: frame.vtReplacementBytes(), encoding: .utf8)
+    let expected = [
+        "\u{1B}[?2026h\u{1B}[0$}\u{1B}[>m\u{1B}[r\u{1B}[?69l\u{1B}[?5W\u{1B}[?1049l",
+        "\u{1B}]8;;\u{1B}\\",
+        "\u{1B}[1\"q\u{1B}[0\"q\u{1B}[999<u\u{1B}[0;1=u\u{0F}\u{1B}(B\u{1B})B\u{1B}*B\u{1B}+B",
+        "\u{1B}]110\u{1B}\\\u{1B}]111\u{1B}\\\u{1B}]112\u{1B}\\",
+        "\u{1B}[0m",
+        "\u{1B}[H\u{1B}[2J\u{1B}[3J\u{1B}[?1049h",
+        "\u{1B}]8;;\u{1B}\\",
+        "\u{1B}[1\"q\u{1B}[0\"q\u{1B}[999<u\u{1B}[0;1=u\u{0F}\u{1B}(B\u{1B})B\u{1B}*B\u{1B}+B",
+        "\u{1B}[0m",
+        "\u{1B}[H\u{1B}[2J\u{1B}[?1049l\u{1B}[H",
+        "\u{1B}[?7l\u{1B}[?25l\u{1B}[0m",
+        "\u{1B}[0m\u{1B}[1Galpha",
+        "\r\n\u{1B}[0m",
+        "\r\n\u{1B}[0m\u{1B}[1G beta",
+        "\r\n\u{1B}[0m",
+        "\u{1B}[0m\u{1B}[2 q\u{1B}[?25h\u{1B}[3;6H",
+        "\u{1B}7",
+        "\u{1B}[?2026l",
+    ].joined()
+    #expect(actual == expected)
 }
 
 @Test func renderGridDeltaClearsOnlyChangedRows() throws {
